move test_estat files
